Backend/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
pytesseract>=0.3.10
scikit-image>=0.22.0
scipy>=1.11.0
pandas>=2.1.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
